# Conflicts:
#	docs/class_sample_generator-members.html
#	docs/class_sample_generator.html
#	docs/functions.html
#	docs/functions_func.html
#	docs/functions_vars.html
#	docs/genetics_8h_source.html
#	docs/globals.html
#	docs/globals_defs.html
#	docs/globals_vars.html
#	docs/multinormal_8h_source.html
#	docs/normal__gen_8cpp.html
#	docs/normal__gen_8h.html
#	docs/normal__gen_8h_source.html
#	docs/population_8h.html
#	docs/population_8h_source.html
#	docs/routes_8h_source.html
#	docs/search/all_0.js
#	docs/search/all_1.js
#	docs/search/all_10.js
#	docs/search/all_12.js
#	docs/search/all_3.js
#	docs/search/all_7.js
#	docs/search/all_9.js
#	docs/search/all_c.js
#	docs/search/defines_8.js
#	docs/search/functions_1.js
#	docs/search/functions_11.js
#	docs/search/functions_12.js
#	docs/search/functions_13.js
#	docs/search/functions_2.js
#	docs/search/functions_3.js
#	docs/search/functions_5.js
#	docs/search/functions_7.js
#	docs/search/functions_a.js
#	docs/search/functions_f.js
#	docs/search/searchdata.js
#	docs/search/variables_0.js
#	docs/search/variables_5.js
#	htdocs/index.html
#	src/routes-lib/genetics/population.cpp
#	src/routes-lib/genetics/population.h
#	src/routes-lib/normal/multinormal.cpp
#	src/routes-lib/normal/normal_gen.cpp
#	src/routes-lib/normal/normal_gen.h

+++ src/routes-lib/normal/multinormal.cpp
@@ -47,7 +47,7 @@ void MultiNormal::generateRandomSamples(std::vector<Eigen::VectorXf>& out, std::
     int num_workers = samplers.size();
     int work_size = out.size() / num_workers;
 
-    std::vector<std::thread> threads (num_workers);
+    std::vector<std::thread> threads ((size_t)num_workers);
 
     for (int i = 0; i < num_workers; i++) {
